src/main/java/com/tw/bootcamp/bookshop/order/OrderController.java
package com.tw.bootcamp.bookshop.order;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RestController;

@RestController
public class OrderController {

    @Autowired
    private OrderService orderService;

    @PostMapping("/orders")
    public ResponseEntity create(@RequestBody Order order) {
        Order createdOrder = orderService.create(order);
        return new ResponseEntity(new OrderResponse(createdOrder.getId()), HttpStatus.CREATED);
    }
}
